refactor: make interfaces prettier
